Lachlan's last commit.
- Changed the win screen to display time as an INT instead of a float with decimal places; it now is rounded.

===Thank you GitHub + Team, till next time <3  ===

+++ Lamb-To-The-Slaughter/Assets/Scripts/UI/winScreen.cs
@@ -5,7 +5,7 @@ public class winScreen : MonoBehaviour //Lachlan
 {
     //Timer value and TXT object
     public TMP_Text timerTxt;
-    public float timerValue;
+    public int timerValue;
 
     //to allow the pause screen to be destoryed
     public GameObject pauseScreen;
